+++ history.md
@@ -43,7 +43,7 @@ node starsky-tools/build-tools/app-version-update.js
 - change default option in thumbnailer cli to scan directories to enabled
 - remove newtonsoft.json references
 
-# version 0.4.8 _(Unreleased)_ - 2021-03-??
+# version 0.4.8 _(Unreleased)_ - 2021-05-??
 - [x]   (Changed) _Front-end_ Make Archive UI more white & Dark mode UI fixes (PR #358)
 - [x]   (Added) _Front-end_  In select mode press delete key should move to trash in archive & search view (PR #360 / Issue #357)
 - [x]   (Fixed)  _Back-end_ Remove check if byte size is the same for example when updating colorClass is the same (PR #362)
@@ -53,8 +53,9 @@ node starsky-tools/build-tools/app-version-update.js
 - [x]   (Added) _Back-end_ Added logs to Application Insights when key is configured (PR #363)
 - [x]   (Added) _Back-end_ Add manual sync for new api (/api/synchronize) and update via websockets (PR #363)
 - [x]   (Added) _Front-end_ Change manual sync to new api (/api/synchronize) (PR #363)
+- [x]   (Fixed) _Front-end_ Force sync is not endless loading after socket update (Issue #371 & PR #375)
 
-# version 0.4.7 - 2021-03-11
+# version 0.4.7 - 2021-04-11
 - [x]   (Changed) _Back-end_  add cache for health check and timeout for 10 seconds on health calls (PR #332)
 - [x]   (Fixed) _Front-end_ Zoom function for mobile (DetailView) (PR #327/ Issue #317)
 - [x]   (Added) _Front-end_ Keyboard shortcuts Cmd + = and Cmd + - (DetailView) (PR #327/ Issue #317)